modernize

+++ src/SkunkLab.Modbus/Messaging/Crc.cs
@@ -1,7 +1,7 @@
 ﻿namespace SkunkLab.Modbus.Messaging
 {
     public class Crc
-    {       
+    {
 
         public static byte[] Compute(byte[] buffer)
         {
@@ -9,7 +9,7 @@
 
             for (int pos = 0; pos < buffer.Length; pos++)
             {
-                crc ^= (ushort)buffer[pos];          // XOR byte into least sig. byte of crc
+                crc ^= buffer[pos];          // XOR byte into least sig. byte of crc
 
                 for (ushort i = 8; i != 0; i--)
                 {    // Loop over each bit
@@ -22,7 +22,7 @@
                         crc >>= 1;                    // Just shift right
                 }
             }
-           
+
             return new byte[2] { (byte)(crc & 0x00FF), (byte)((crc >> 8) & 0x00FF) };
         }
     }
